ingestion/requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-genai==0.3.0
beautifulsoup4==4.12.3
curl_cffi==0.7.1
numpy>=1.26.0
scikit-learn>=1.4.0
